tests and tweaks
