+++ lib/cog/commands/rule/info.ex
@@ -27,8 +27,17 @@ defmodule Cog.Commands.Rule.Info do
 
   rule "when command is #{Cog.Util.Misc.embedded_bundle}:rule-info must have #{Cog.Util.Misc.embedded_bundle}:manage_commands"
 
-  def handle_message(req = %{args: [id]}, state) when is_binary(id) do
-    result = if Cog.UUID.is_uuid?(id) do
+  def handle_message(req, state) do
+    case info(req) do
+      {:ok, rule} ->
+        {:reply, req.reply_to, "rule-info", rule, state}
+      {:error, error} ->
+        {:error, req.reply_to, RuleCommand.error(error), state}
+    end
+  end
+
+  defp info(%{args: [id]}) when is_binary(id) do
+    if Cog.UUID.is_uuid?(id) do
       case Rules.rule(id) do
         %Rule{}=rule ->
           {:ok, Cog.V1.RuleView.render("show.json", %{rule: rule})[:rule]}
@@ -38,19 +47,12 @@ defmodule Cog.Commands.Rule.Info do
     else
       {:error, {:rule_uuid_invalid, id}}
     end
-
-    case result do
-      {:ok, rule} ->
-        {:reply, req.reply_to, "rule-info", rule, state}
-      {:error, error} ->
-        {:error, req.reply_to, RuleCommand.error(error), state}
-    end
   end
-  def info(%{args: [_]}, _state),
+  defp info(%{args: [_]}),
     do: {:error, :wrong_type}
-  def info(%{args: []}, _state),
+  defp info(%{args: []}),
     do: {:error, {:not_enough_args, 1}}
-  def info(_req, _state),
+  defp info(_req),
     do: {:error, {:too_many_args, 1}}
 
 end
